Add modelcache internally
